requirements.txt
pandas
plotly-express
seaborn
matplotlib
numpy
streamlit